// frontend/src/gen/inventory_iface/v1/stock_pb.ts
export enum MovementType {
  /**
   * @generated from enum value: MOVEMENT_TYPE_UNSPECIFIED = 0;
   */
  UNSPECIFIED = 0,

  /**
   * @generated from enum value: MOVEMENT_TYPE_PURCHASE = 1;
   */
  PURCHASE = 1,

  /**
   * @generated from enum value: MOVEMENT_TYPE_SALE = 2;
   */
  SALE = 2,

  /**
   * @generated from enum value: MOVEMENT_TYPE_ADJUSTMENT = 3;
   */
  ADJUSTMENT = 3,

  /**
   * @generated from enum value: MOVEMENT_TYPE_WRITE_OFF = 4;
   */
  WRITE_OFF = 4,
}
// ...
export class StockMovement extends Message<StockMovement> {
  /**
   * @generated from field: string id = 1;
   */
  id = "";

  /**
   * @generated from field: string batch_id = 2;
   */
  batchId = "";

  /**
   * @generated from field: int32 qty = 3;
   */
  qty = 0;

  /**
   * @generated from field: inventory_iface.v1.MovementType type = 4;
   */
  type = MovementType.UNSPECIFIED;

  /**
   * @generated from field: string reason = 5;
   */
  reason = "";

  /**
   * @generated from field: string user_id = 6;
   */
  userId = "";

  /**
   * @generated from field: int64 created_at = 7;
   */
  createdAt = protoInt64.zero;

  constructor(data?: PartialMessage<StockMovement>) {
    super();
    proto3.util.initPartial(data, this);
  }

  static readonly runtime: typeof proto3 = proto3;
  static readonly typeName = "inventory_iface.v1.StockMovement";
  static readonly fields: FieldList = proto3.util.newFieldList(() => [
    { no: 1, name: "id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 2, name: "batch_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 3, name: "qty", kind: "scalar", T: 5 /* ScalarType.INT32 */ },
    { no: 4, name: "type", kind: "enum", T: proto3.getEnumType(MovementType) },
    { no: 5, name: "reason", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 6, name: "user_id", kind: "scalar", T: 9 /* ScalarType.STRING */ },
    { no: 7, name: "created_at", kind: "scalar", T: 3 /* ScalarType.INT64 */ },
  ]);

  static fromBinary(bytes: Uint8Array, options?: Partial<BinaryReadOptions>): StockMovement {
    return new StockMovement().fromBinary(bytes, options);
  }

  static fromJson(jsonValue: JsonValue, options?: Partial<JsonReadOptions>): StockMovement {
    return new StockMovement().fromJson(jsonValue, options);
  }

  static fromJsonString(jsonString: string, options?: Partial<JsonReadOptions>): StockMovement {
    return new StockMovement().fromJsonString(jsonString, options);
  }

  static equals(a: StockMovement | PlainMessage<StockMovement> | undefined, b: StockMovement | PlainMessage<StockMovement> | undefined): boolean {
    return proto3.util.equals(StockMovement, a, b);
  }
}